requirements.txt
opencv-python
numpy
fer
matplotlib
streamlit
Pillow
scikit-image
opencv-contrib-python